+++ src/time_split/__main__.py
@@ -1,0 +1,9 @@
+"""Alias of :mod:`time_split.cli`.
+
+As per, https://docs.python.org/3.13/library/__main__.html#id1,
+this file doesn't use a ``if __name__ == "__main__"``-block.
+"""
+
+from .cli import main
+
+main()