+++ src/time_split/app/__init__.py
@@ -1,6 +1,7 @@
-"""Supporting functions and classes for the Streamlit companion app.
+"""Supporting functions for the Streamlit companion app.
 
-This is the public API of the Streamlit companion app.
+The app may be installed by running ``pip install time-split[app]``. The application API is *not* stable, but will be
+available under the ``time_split_app`` namespace. Repo: https://github.com/rsundqvist/time-split-app/.
 """
 
 from ._create_explorer_link import create_explorer_link